Add files via upload
In this experiment we aim to predict the pricing of the Paris houses using the independent variables. The prediction model to be used is the linear regression model. This is a regression task. There is a total of 17 different variables in this experiment, and the independent variables are squareMeters, numberOfRooms, floors, cityCode, cityPartRange, numPrevOwners, made, isNewBuilt, hasStormProtector, basement, attic, garage, hasStorageRoom, hasGuestRoom, category, and PoolAndYard. The dependent variable is the price. The public dataset link is: https://www.kaggle.com/datasets/aleshagavrilov/parishousing

QNA

1)What are the data types? (Only numeric and categorical)

2)Are there missing values?

3)What are the likely distributions of the numeric variables?

4)Which independent variables are useful to predict a target (dependent variable)?

5)Which independent variables have missing data? How much?

6)Do the training and test sets have the same data?

7)In the predictor variables independent of all the other predictor variables?

8)Which predictor variables are the most important?

9)Do the ranges of the predictor variables make sense?

10)What are the distributions of the predictor variables?

11)Remove outliers and keep outliers (does if have an effect of the final predictive model)?

12)Remove 1%, 5%, and 10% of your data randomly and impute the values back using at least 3 imputation methods. How well did the methods recover the missing values?